feat: TOP-28 LocationService 생성하여 DataSource 로직 이전
