+++ data/src/main/java/com/tht/tht/data/remote/service/LocationService.kt
@@ -1,0 +1,7 @@
+package com.tht.tht.data.remote.service
+
+import com.tht.tht.data.remote.response.location.LocationResponse
+
+interface LocationService {
+    suspend fun fetchCurrentLocation(): LocationResponse
+}
